+++ module/actor/character-sheet-gen8.js
@@ -33,7 +33,6 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 		}
 
 		data['origins'] = this.actor.origins;
-		debug(data.origins, this.actor);
 
 		return data;
 	}
@@ -135,6 +134,18 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 			effect.sheet.render(true);
 		});
 
+		// Disable Effect
+		html.find('.effect-suspend').click(async (ev) => {
+			ev.preventDefault();
+			const button = ev.currentTarget;
+			const effectId = button.dataset.id;
+			const effect = this.actor.effects.get(effectId);
+			const effectData = duplicate(effect.data);
+			effectData.disabled = !effectData.disabled;
+			await effect.update(effectData);
+			this.render(false);
+		});
+
 		// Delete Inventory Item
 		html.find('.item-delete').click((ev) => {
 			const li = $(ev.currentTarget).parents('.item');
@@ -177,6 +188,16 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 			autoFocus: true,
 			minLength: 1
 		});
+
+		Hooks.on("preCreateOwnedItem", (actor, itemData, options, sender) => {
+			if(actor.id !== this.actor.id || actor.data.type !== "character" || itemData.type !== "dexentry") return;
+			
+			const item = actor.items.getName(itemData.name)
+			if(item) {
+				//if(!item.data.owned) item.update({"data.owned": true});
+				return false;
+			}
+		})
 	}
 
 	_onDragItemStart(event) {}
@@ -208,6 +229,9 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 			// Finally, create the effect!
 			debug("Created new effect",itemData);
 			return this.actor.createEmbeddedEntity(itemData.type, itemData);
+		}
+		if(itemData.type === "dexentry") {
+
 		}
 
 		// Finally, create the item!