+++ module/ptu.js
@@ -50,7 +50,7 @@ export let log = (...args) => console.log("FVTT PTU | ", ...args);
 export let warn = (...args) => console.warn("FVTT PTU | ", ...args);
 export let error = (...args) => console.error("FVTT PTU | ", ...args)
 
-export const LATEST_VERSION = "1.2.16";
+export const LATEST_VERSION = "1.2.17";
 
 function registerSheets() {
   // Register sheet application classes